+++ chag/__init__.py
@@ -1,3 +1,6 @@
+from .changelog import Changelog, Entry
+from . import git
+
 __author__ = 'Michael Dowling'
 __email__ = 'mtdowling@gmail.com'
-__version__ = '0.3.0'
+__version__ = '0.5.0'